+++ tests/smoke/test_smoke.py
@@ -34,6 +34,13 @@ def settings():
 
     settings["cbf_ec_mount_path"] = CBF_EC_MOUNT_PATH
 
+    # Get SPFRX IPs
+    spfrx_hw_config_relative_path = "resources/spfrx/spfrx_hw_config.yaml"
+    with open(spfrx_hw_config_relative_path, "r") as f:
+        spfrx_talon_board_ips = yaml.safe_load(f)["talon_board"]
+
+    settings["spfrx_talon_board_ips"] = spfrx_talon_board_ips
+
     return settings
 
 
@@ -66,13 +73,9 @@ def test_qspi_bitstream_compatibility(settings):
 
     # Get CBF Engineering console version and expected fpga bitstream version
     umbrella_chart_relative_path = "charts/ska-mid/Chart.yaml"
-    with open(umbrella_chart_relative_path, "r") as f:
-        sut_chart = yaml.safe_load(f)
-
-        for dependency in sut_chart["dependencies"]:
-            if dependency["name"] == "ska-mid-cbf-engineering-console":
-                cbf_engineering_console_version = dependency["version"]
-                break
+    cbf_engineering_console_version = get_chart_dependency_version(
+        umbrella_chart_relative_path, "ska-mid-cbf-engineering-console"
+    )
 
     fpga_bitstream_version = TalonBoardCommandExecutor.get_fpga_bitstream_version(
         cbf_engineering_console_version
@@ -138,3 +141,78 @@ def test_qspi_bitstream_compatibility(settings):
             f"Checksum computed from CBF EC: {bitstream_checksum}, "
             f"Checksum reported on Talon board slot {slot_number}: {loaded_bitstream_checksum}. "
         )
+
+
+def test_spfrx_qspi_bitstream_compatibility(settings):
+    """Check QSPI bitstream version for SPFRX Talon Boards.
+
+    :param settings: Smoke test configuration
+    :type settings: Dict
+    """
+    spfrx_talon_board_ips = settings["spfrx_talon_board_ips"]
+
+    # Get SPFRx console version from charts/ska-mid/Chart.yaml
+    umbrella_chart_relative_path = "charts/ska-mid/Chart.yaml"
+    spfrx_talondx_console_version = get_chart_dependency_version(
+        umbrella_chart_relative_path, "ska-mid-dish-spfrx-talondx-console"
+    )
+
+    spfrx_bitstream_version = TalonBoardCommandExecutor.get_spfrx_bitstream_version(
+        spfrx_talondx_console_version
+    )
+    logger.info(f"FPGA bitstream version: {spfrx_bitstream_version}")
+
+    # Get actual bitstream version from talonboards and compare with expected version
+    user = "root"
+    for talon_board, ip in spfrx_talon_board_ips.items():
+        talon_board_command_executor = TalonBoardCommandExecutor(ip, user)
+        command_result = talon_board_command_executor.execute_command(
+            talon_board, "qspi_version_check"
+        )
+        if command_result is None:
+            pytest.fail(f"Failed to execute command successfully on Talon board {talon_board}")
+
+        # Get currently loaded slot
+        slot_number = talon_board_command_executor.get_currently_loaded_slot(command_result)
+        if slot_number is None:
+            pytest.fail(f"Failed to get currently used slot on Talon board {talon_board}")
+
+        logger.info(f"Talon {talon_board} loaded slot: {slot_number}")
+
+        # Get bitstream version at slot
+        loaded_bitstream_version = talon_board_command_executor.get_loaded_bitstream_version(
+            slot_number, command_result
+        )
+        if loaded_bitstream_version is None:
+            pytest.fail(f"Failed to get bitstream version on Talon board {talon_board}")
+        logger.info(f"SPFRx Talon {talon_board} bitstream version: {loaded_bitstream_version}")
+
+        # Check compatibility
+        bitstream_compatible = TalonBoardCommandExecutor.check_spfrx_bitstream_compatibility(
+            loaded_bitstream_version,
+            spfrx_bitstream_version,
+        )
+
+        assert bitstream_compatible
+
+
+def get_chart_dependency_version(umbrella_chart_relative_path: str, dependency_name: str):
+    """Get dependency version from umbrella chart.
+
+    :param umbrella_chart_relative_path: Path (relative to repo root) to the umbrella chart
+    :type umbrella_chart_relative_path: str
+    :param dependency_name: Name of the dependency to look for
+    :type dependency_name: str
+    :return: Version of the dependency, returns None if not found
+    :rtype: str | None
+    """
+    version = None
+    with open(umbrella_chart_relative_path, "r") as f:
+        chart = yaml.safe_load(f)
+
+        for dependency in chart["dependencies"]:
+            if dependency["name"] == dependency_name:
+                version = dependency["version"]
+                break
+
+    return version